Remove the variable-replace stuff from CDK to a separate script
