+++ src/js/grapl-cdk/deploy_all.sh
@@ -1,7 +1,8 @@
 #!/usr/bin/env bash
 npm run build &&
-cdk deploy --require-approval=never Grapl && \
+cdk deploy --require-approval=never --outputs-file=./cdk-output.json Grapl && \
 rm -rf ./edge_ux_package && \
+npm run create_edge_ux_package && \
 cdk synth && \
 cdk deploy --require-approval=never EngagementUX && \
